+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The last line of some popup help documentation is no longer cut off.
 - Windows compatibility issues fixed for `file-author` and `send-signal`.
 - Stack trace is now printed on error in non-interactive mode.
+- `class-supers` now supports structs and any other class.
+- `class-precedence` now supports structs and any other class.
+
 ### Added
 - Added `-b` command line option with binding to `$@` and `$<n>` where <n> is the argument position.
 - The `bag-scan` function to scan a `bag` and call a function on each node in the data.